requirements.txt
fastapi==0.135.1
groq==1.1.1
Jinja2==3.1.6
pydantic==2.12.5
python-dotenv==1.2.2
uvicorn==0.42.0
